# openapi.yaml
openapi: 3.0.3
info:
  title: MVTMS REST API
  description: |
    OpenAPI specification for the MVTMS REST API v1. Authentication is via
    a signed Bearer token (`Authorization: Bearer <token>`) or an existing
    session cookie. The token endpoints are public; all other `/api/v1/*`
    routes require a valid principal plus Casbin RBAC permissions.
  version: "1.0.0"
  contact:
    name: MVTMS Support
    url: https://avandab.com
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/bookings:
    get:
      tags: [Bookings]
      summary: List bookings
      description: Requires `bookings:read` permission.
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SearchParam'
        - $ref: '#/components/parameters/StatusParam'
      responses:
        '200':
          description: Paginated list of bookings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Bookings]
      summary: Create a booking
      description: Requires `bookings:create` permission.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
      responses:
        '201':
          description: Booking created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IDResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/bookings/{id}:
    get:
      tags: [Bookings]
      summary: Get a booking
      description: Requires `bookings:read` permission.
      parameters:
        - $ref: '#/components/parameters/IDParam'
      responses:
        '200':
          description: Booking details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Bookings]
      summary: Update a booking
      description: Requires `bookings:update` permission.
      parameters:
        - $ref: '#/components/parameters/IDParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBookingRequest'
      responses:
        '200':
          description: Booking updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    delete:
      tags: [Bookings]
      summary: Delete a booking
      description: Requires `bookings:delete` permission.
      parameters:
        - $ref: '#/components/parameters/IDParam'
      responses:
        '204':
          description: Booking deleted.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/bookings/{id}/confirm:
    post:
      tags: [Bookings]
      summary: Confirm a booking
      description: Requires `bookings:approve` permission.
      parameters:
        - $ref: '#/components/parameters/IDParam'
      responses:
        '200':
          description: Booking confirmed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/bookings/{id}/cancel:
    post:
      tags: [Bookings]
      summary: Cancel a booking
      description: Requires `bookings:cancel` permission.
      parameters:
        - $ref: '#/components/parameters/IDParam'
      responses:
        '200':
          description: Booking cancelled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/bookings/{id}/complete:
    post:
      tags: [Bookings]
      summary: Complete a booking
      description: Requires `bookings:update` permission.
      parameters:
        - $ref: '#/components/parameters/IDParam'
      responses:
        '200':
          description: Booking completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Signed API token returned by `POST /api/v1/auth/token`.
  parameters:
    IDParam:
      name: id
      in: path
      required: true
      schema:
        type: string
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
      description: Page number (1-based).
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 10
      description: Items per page.
    SearchParam:
      name: search
      in: query
      schema:
        type: string
      description: Free-text search filter.
    StatusParam:
      name: status
      in: query
      schema:
        type: string
      description: Status filter.
  schemas:
...
    IDResponse:
      type: object
      properties:
        id:
          type: string
    StatusResponse:
      type: object
      properties:
        status:
          type: string
    CreateBookingRequest:
      type: object
      required: [customer_id, route_id, vehicle_type, passengers, price]
      properties:
        customer_id:
          type: string
        route_id:
          type: string
        pickup_date:
          type: string
          example: "2026-08-20"
          description: |
            Accepts `YYYY-MM-DD`, RFC3339, `YYYY-MM-DD HH:MM`, or
            `YYYY-MM-DD HH:MM:SS`.
        vehicle_type:
          type: string
        passengers:
          type: integer
          minimum: 1
        cargo_weight:
          type: number
          nullable: true
        price:
          type: number
          minimum: 0
        notes:
          type: string
    UpdateBookingRequest:
      allOf:
        - $ref: '#/components/schemas/CreateBookingRequest'
    Booking:
      type: object
      properties:
        id:
          type: string
        booking_number:
          type: string
        customer_id:
          type: string
        route_id:
          type: string
        pickup_date:
          type: string
          format: date-time
        vehicle_type:
          type: string
        passengers:
          type: integer
        cargo_weight:
          type: number
          nullable: true
        price:
          type: number
        notes:
          type: string
        status:
          type: string
          enum: [draft, pending, confirmed, cancelled, completed]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    BookingList:
      type: object
      properties:
        bookings:
          type: array
          items:
            $ref: '#/components/schemas/Booking'
        total:
          type: integer
          format: int64
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
  responses:
    BadRequest:
      description: Bad request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required or invalid credentials.
      headers:
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient RBAC permissions.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'